Collect test fixtures in .testing
Remove tests/conftest.py.
